requirements.txt
crewai
crewai_tools
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
oauth2client
python-dotenv
flask
langchain-openai